fix: set lower bound of 1 for games played
If a league has bye weeks, those team's `team.outcomes` will be populated with a `'U'` and not included in the denominator when calculating the win percentage. However, if a team had a bye in week 1, then their total games played will sum to 0 and raise a `ZeroDivisionError: float division by zero`. This bounds the denominator at 1 so that a 0.0 `win_pct` is calculated instead.